First Commit
The main goal of this project is to analyse extensive data on the production of coffee, processing techniques, sensory analysis, and defect occurrences. Through the use of this data, the project hopes to unearth insights that will benefit all parties involved in the coffee business, from growers and producers to roasters and retailers, in order to enhance coffee quality, optimize production procedures, and eventually provide customers with an improved coffee experience.